Random, randint and choice usage example
First, this program initializes a class called "Die" and gives the die 7 sides. Then it simulates rolling the die and prints which side it landed on. 
Next, it creates a tuple of characters, an empty list and a counter variable. It randomly populates the list with characters from the tuple. Next, it initiates a list called my_ticket and uses a while loop to compare my_ticket with the randomly populated list. It regenerates the randomly populated list until it matches my_ticket, and records how many tries it took, and prints them at the end. This program uses the random API with randint and choice.