Dictionary and enumerate usage.
This program opens a file called worldserieswinners.txt and enumerates. It creates two variables, one for team name and one for year, and it start the year at 1903. Then, it reads each team name into the team variable and each year into the year variable. It then adds each of these values as key value pairs to a dictionary; it adds the team as a key, but it creates a list of values for the value associated with each time and appends each year a team has won besides the initial year that team won. If the team is not in the list it prints a message saying that team is not in the list.