Class and Dictionary usage
Write a class Movie. It should have attributes for the following data:○The movie’s name.○A dictionary initialized with
5 key/value pairs: The keys should be the numbers 1 through 5, and each value should be initialized to 0. The values
signify how many reviews the movie has received with that many stars. i.e., If key 1 is paired with the value 3, that
means three users have given the movie 1 star.○An int variable num_reviews, which tracks how many reviews the film has
received. ○A constructor that accepts the movie’s name as an argument, sets num_reviews to 0, and initializes the
dictionary’s key/value pairs

It should have the following methods:○A property setter and getter for the movie’s name.○add_review, which can only be
passed the values 1 through 5. As an example, if ‘2’ is passed, that means a review of 2 stars is being added:
increment the value associated with the key ‘2’ by one.○