Pickle usage. Class usage: Getters & Setters.
This code uses pickling and classes with getters and setters. Pickling is serializing an object. Pickling is used for storage or data transport. This program pickles a dictionary and allows the user to add a contact to it (which is pickled in the process), or retrieve a contact from it (which is unpickled in the process).